refactor(feature/tratamento-de-erros-refatoracao): organizacao dos packages
