+++ AluGames/src/main/kotlin/br/com/luizlemetech/alugames/modelo/InfoApiShark.kt
@@ -1,3 +1,3 @@
-package com.br.luizleme.tech.alugames.dominio
+package br.com.luizlemetech.alugames.modelo
 
 data class InfoApiShark(val title:String, val thumb:String)
